Add integration test and change define test port
This commit define the Jest test configuration to use port 4000 instead of the default port 3000, to avoid conflicts with the application server. This is necessary because the application server is automatically started when the script "npm run dev" is launched causes conflicts when running Jest tests that also use port 3000. By changing the test port to 4000, we ensure that the tests can run without any issues, meanwhile the server is live. This change will not affect the application server and it will continue to run on port 3000 as before.